Rename schema.py to table_ddl.py
The name 'schema' was overloaded: it collided with the DomainSchema
class and the Postgres schema concept. 'table_ddl' better describes the
module's role of defining case-type tables and emitting their DDL.
